+++ packages/k6/entry.sh
@@ -40,7 +40,7 @@ if [ ! -z ${PR_NUMBER+x} ]; then
   TEST_OUTPUT_ASSERTIONS_FAILED=$(cat ~/output.log | grep "Failed: " | sort | uniq)
   TEST_OUTPUT_ASSERTIONS_LOADED_COUNT=$(cat ~/output.log | grep "Assertion loaded: " | sort | uniq | wc -l)
   TEST_OUTPUT_ASSERTIONS_COUNT=$(cat ~/output.log | grep "Assertion applied: " | sort | uniq | wc -l)
-  TEST_OUTPUT_SAMPLE=$(cat ~/output.log | grep -e "request: " -e "DEBUG: " | tail -n 200)
+  TEST_OUTPUT_SAMPLE=$(cat ~/output.log | tail -n 1000)
   TEST_OUTPUT_PARAM_NUM=$(sed 's/^request: \(.*\) response.*/\1/' ~/output.log | sort | uniq | wc -l)
   if [ "$TEST_OUTPUT_PARAM_NUM" -lt 5 ]; then
     TEST_OUTPUT_MESSAGE=":warning: Only $TEST_OUTPUT_PARAM_NUM unique input parameter sets. Update test-payload.json to increase the coverage. "