+++ packages/sources/bx-digital/src/endpoint/price.ts
@@ -1,0 +1,33 @@
+import { AdapterEndpoint } from '@chainlink/external-adapter-framework/adapter'
+import { InputParameters } from '@chainlink/external-adapter-framework/validation'
+import { SingleNumberResultResponse } from '@chainlink/external-adapter-framework/util'
+import { config } from '../config'
+import { httpTransport } from '../transport/price'
+
+export const inputParameters = new InputParameters(
+  {
+    securityId: {
+      required: true,
+      type: 'string',
+      description: 'ID of the security to report price on',
+    },
+  },
+  [
+    {
+      securityId: 'CH0012032048',
+    },
+  ],
+)
+
+export type BaseEndpointTypes = {
+  Parameters: typeof inputParameters.definition
+  Response: SingleNumberResultResponse
+  Settings: typeof config.settings
+}
+
+export const endpoint = new AdapterEndpoint({
+  name: 'price',
+  aliases: [],
+  transport: httpTransport,
+  inputParameters,
+})